+++ tests/test_async.py
@@ -101,7 +101,7 @@ def test_autorun(
     async def _(mirrored_value: int) -> None:
         if mirrored_value < INCREMENTS:
             return
-        await asyncio.sleep(0.1)
+        await asyncio.sleep(0.001)
         store.dispatch(FinishAction())
 
     dispatch_actions(store)
@@ -124,7 +124,7 @@ def test_autorun_autoawait(store: StoreType) -> None:
         elif value < INCREMENTS:
             store.dispatch(IncrementAction())
         else:
-            await asyncio.sleep(0.1)
+            await asyncio.sleep(0.001)
             store.dispatch(FinishAction())
 
 
@@ -139,6 +139,7 @@ def test_autorun_non_autoawait(store: StoreType) -> None:
     @store.autorun(lambda state: (state.value, state.mirrored_value))
     async def _(values: tuple[int, int]) -> None:
         value, mirrored_value = values
+        await asyncio.sleep(0.001)
         if mirrored_value != value:
             sync_mirror_returned_value = sync_mirror()
             assert 'awaited=False' in str(sync_mirror_returned_value)
@@ -148,7 +149,6 @@ def test_autorun_non_autoawait(store: StoreType) -> None:
         elif value < INCREMENTS:
             store.dispatch(IncrementAction())
         else:
-            await asyncio.sleep(0.1)
             store.dispatch(FinishAction())
 
 
@@ -165,7 +165,7 @@ def test_autorun_default_value(store: StoreType) -> None:
     async def _(mirrored_value: int) -> None:
         if mirrored_value < INCREMENTS:
             return
-        await asyncio.sleep(0.1)
+        await asyncio.sleep(0.001)
         store.dispatch(FinishAction())
 
     dispatch_actions(store)
@@ -184,7 +184,7 @@ def test_view(store: StoreType) -> None:
         assert await doubled() == value * 2
         for _ in range(10):
             await doubled()
-        await asyncio.sleep(0.01)
+        await asyncio.sleep(0.001)
         if value < INCREMENTS:
             store.dispatch(IncrementAction())
         else:
@@ -209,7 +209,7 @@ def test_view_await(store: StoreType) -> None:
         calls_length = len(calls)
         assert await doubled() == value * 2
         assert len(calls) == calls_length + 1
-        await asyncio.sleep(0.01)
+        await asyncio.sleep(0.001)
 
         if value < INCREMENTS:
             store.dispatch(IncrementAction())
@@ -230,7 +230,7 @@ def test_view_with_args(store: StoreType) -> None:
     async def _(value: int) -> None:
         assert await multiplied(factor=2) == value * 2
         assert await multiplied(factor=3) == value * 3
-        await asyncio.sleep(0.01)
+        await asyncio.sleep(0.001)
         if value < INCREMENTS:
             store.dispatch(IncrementAction())
         else:
@@ -251,7 +251,7 @@ def test_view_with_default_value(store: StoreType) -> None:
     @store.autorun(lambda state: state.value)
     async def _(value: int) -> None:
         assert await doubled() == value * 2
-        await asyncio.sleep(0.01)
+        await asyncio.sleep(0.001)
         if value < INCREMENTS:
             store.dispatch(IncrementAction())
         else:
@@ -265,7 +265,7 @@ def test_subscription(store: StoreType) -> None:
     async def render(state: StateType) -> None:
         if state.value == INCREMENTS:
             unsubscribe()
-            await asyncio.sleep(0.1)
+            await asyncio.sleep(0.001)
             store.dispatch(FinishAction())
 
     unsubscribe = store._subscribe(render)  # noqa: SLF001
@@ -277,7 +277,7 @@ def test_event_subscription(store: StoreType) -> None:
     async def handler(event: IncrementEvent) -> None:
         if event.post_value == INCREMENTS:
             unsubscribe()
-            await asyncio.sleep(0.1)
+            await asyncio.sleep(0.001)
             store.dispatch(FinishAction())
 
     unsubscribe = store.subscribe_event(IncrementEvent, handler)