+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - chore: add badges in `README.md` and classifiers in `pyproject.toml`
 - refactor: move the common code for manipulating the signature of the wrapped functions in `WithStore` and `Autorun` to a utility function
+- feat: support `with_state` to be applied to methods of classes, not just functions
 
 ## Version 0.23.0